添加groupsize约束

+++ mc2/allto_all_matmul/op_host/allto_all_matmul_infershape.cpp
@@ -42,15 +42,15 @@ constexpr int64_t NUM_MINUS_ONE = -1;
 constexpr int64_t NUM_MINUS_TWO = -2;
 constexpr int64_t OUTPUT_INFER_SHAPE = 2;
 static const char* INNER_DEBUG = "MC2: AlltoAllMatmul InferShape Debug";
-const std::set<int> SUPPORT_RANK_NUM{2, 4, 8, 16};
+const std::set<int64_t> SUPPORT_RANK_NUM{2, 4, 8, 16};
 
 struct AlltoAllMatmulShapeInfo {
-    uint64_t output_dim;
-    uint64_t rankNum;
-    uint64_t m;
-    uint64_t n;
-    uint64_t k1;
-    uint64_t k2;
+    int64_t output_dim;
+    int64_t rankNum;
+    int64_t m;
+    int64_t n;
+    int64_t k1;
+    int64_t k2;
 };
 
 /**
@@ -102,7 +102,7 @@ static ge::graphStatus CheckShapeForAlltoAllMatmul(const gert::InferShapeContext
 static ge::graphStatus CheckRankDim(gert::InferShapeContext* context, AlltoAllMatmulShapeInfo& shape)
 {
     const auto attrs = context->GetAttrs();
-    const int* rankDim = attrs->GetAttrPointer<int>(INDEX_ATTR_WORLD_SIZE);
+    const int64_t* rankDim = attrs->GetAttrPointer<int64_t>(INDEX_ATTR_WORLD_SIZE);
     OPS_CHECK(rankDim == nullptr,
         CUBE_INNER_ERR_REPORT(context->GetNodeName(), "Invalid rank number %zu in allto all matmul.", *rankDim),
         return ge::GRAPH_FAILED);
@@ -140,8 +140,8 @@ static ge::graphStatus InferShapeAlltoAllMatmul(gert::InferShapeContext* context
     all2all_out->SetDimNum(OUTPUT_INFER_SHAPE);
     if (all2all_out_flag) {
         if (shape.m != NUM_MINUS_ONE) {
-            uint64_t all2all_out_first_dim = CeilDiv(shape.m, shape.rankNum);
-            uint64_t all2all_out_second_dim = shape.k1 * shape.rankNum;
+            int64_t all2all_out_first_dim = CeilDiv(shape.m, shape.rankNum);
+            int64_t all2all_out_second_dim = shape.k1 * shape.rankNum;
             all2all_out->SetDim(0U, all2all_out_first_dim);
             all2all_out->SetDim(1U, all2all_out_second_dim);
             OP_LOGI(
@@ -154,8 +154,8 @@ static ge::graphStatus InferShapeAlltoAllMatmul(gert::InferShapeContext* context
         shape_out->SetDim(0U, shape.m);
         shape_out->SetDim(1U, shape.n);
     } else {
-        uint64_t out_first_dim = CeilDiv(shape.m, shape.rankNum);
-        uint64_t out_second_dim = shape.n;
+        int64_t out_first_dim = CeilDiv(shape.m, shape.rankNum);
+        int64_t out_second_dim = shape.n;
         shape_out->SetDim(0U, out_first_dim);
         shape_out->SetDim(1U, out_second_dim);
         OP_LOGI(INNER_DEBUG, "allto all matmul output shape after infer shape, m: %ld n: %ld.", out_first_dim, out_second_dim);
@@ -174,8 +174,8 @@ static ge::graphStatus InferDataTypeAlltoAllMatmul(gert::InferDataTypeContext* c
     OP_LOGD(INNER_DEBUG, "Start to infer datatype of allto all matmul.");
     const auto attrs = context->GetAttrs();
     OPS_CHECK_NULL_WITH_CONTEXT(context, attrs);
-    const int* x1_quant_mode = attrs->GetAttrPointer<int>(INDEX_ATTR_X1_QUANT_MODE);
-    const int* x2_quant_mode = attrs->GetAttrPointer<int>(INDEX_ATTR_X2_QUANT_MODE);
+    const int64_t* x1_quant_mode = attrs->GetAttrPointer<int64_t>(INDEX_ATTR_X1_QUANT_MODE);
+    const int64_t* x2_quant_mode = attrs->GetAttrPointer<int64_t>(INDEX_ATTR_X2_QUANT_MODE);
     const int64_t* y_dtypes_ptr = attrs->GetInt(INDEX_ATTR_Y_DTYPE);
     auto y_type = ge::DataType::DT_UNDEFINED;
     ge::DataType x1_type = context->GetInputDataType(INDEX_IN_X1);
